+++ lib/bundler/dependencies.rb
@@ -9,10 +9,13 @@ require 'bundler/dependencies/spec'
 require 'bundler/dependencies/visitor'
 
 require 'bundler/dependencies/cli/command'
+require 'bundler/dependencies/cli/with_gem'
 require 'bundler/dependencies/cli/count'
+require 'bundler/dependencies/cli/find'
 require 'bundler/dependencies/cli/graph'
 
 require 'bundler/dependencies/visitors/shell_tree'
+require 'bundler/dependencies/visitors/paths'
 
 module Bundler
   module Dependencies