+++ lib/bundler/dependencies/cli/with_gem.rb
@@ -1,0 +1,16 @@
+module Bundler
+  module Dependencies
+    class CLI < ::Thor
+      module WithGem
+        def initialize(gem, options)
+          @gem = gem
+          super(options)
+        end
+
+      private
+
+        attr_reader :gem
+      end
+    end
+  end
+end